Added API paging ability
We've added API paging functionality to help alleviate the pain of
{en,de}coding large pieces of JSON data. What this means is that we
will be splitting calls to the API that would return large amounts of
data into multiple, smaller calls.

From a developer perspective the interface should remain the same. The
only difference to keep in mind is that multiple API calls are being made,
and if ANY fail we consider the whole call a failure. This is so we
don't have to worry about returning and handling partial data.